decreased pa_remap to 6192 bytes (97 cpu cachelines)

+++ src/pulsecore/remap.h
@@ -30,7 +30,7 @@ typedef void (*pa_do_remap_func_t) (pa_remap_t *m, void *d, const void *s, unsig
 struct pa_remap {
     pa_sample_format_t format;
     pa_sample_spec i_ss, o_ss;
-    float map_table_f[PA_CHANNELS_MAX][PA_CHANNELS_MAX];
+    _Float16 map_table_f[PA_CHANNELS_MAX][PA_CHANNELS_MAX];
     int32_t map_table_i[PA_CHANNELS_MAX][PA_CHANNELS_MAX];
     pa_do_remap_func_t do_remap;
     void *state; /* optional state information for the remap function */
